+++ sources/ScreenshotAnnotator/Models/ApplicationSettingsV1Dto.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+
 namespace ScreenshotAnnotator.Models;
 
 public class ApplicationSettingsV1Dto
@@ -5,4 +7,6 @@ public class ApplicationSettingsV1Dto
     public bool IsFileBrowserVisible { get; set; } = true;
     public uint SelectedHighlighterColorArgb { get; set; } = 0x64FFFF00;
     public bool EnablePrintScreenHotkey { get; set; } = true;
+    public string? OcrPreferredEngine { get; set; }
+    public List<string> OcrPreferredLanguages { get; set; } = new();
 }